Add version 1 on API endpoints
