--- app/controllers/api/companies_controller.rb
@@ -1,17 +1,0 @@
-module Api
-  class CompaniesController < ApiController
-    def users
-      render json: company.users.active.engineer.as_json(only: %i[email name github office_id])
-    end
-
-    def offices
-      render json: company.offices
-    end
-
-    private
-
-    def company
-      @company ||= current_user.company
-    end
-  end
-end
