+++ app/controllers/api/v1/companies_controller.rb
@@ -1,0 +1,19 @@
+module Api
+  module V1
+    class CompaniesController < ApiController
+      def users
+        render json: company.users.active.engineer.as_json(only: %i[email name github office_id])
+      end
+
+      def offices
+        render json: company.offices
+      end
+
+      private
+
+      def company
+        @company ||= current_user.company
+      end
+    end
+  end
+end
